chore(tazjin/nixos): update initial hashed password
I still have the issue where after reboots I can't log in to my
machines without going through a horrible dance with booting a NixOS
installer and so on.

I suspect this has somethign to do with this initial hashed password
set here, but I was unable to verify what password I've actually set
up there, so I'm resetting it to a known string to verify my theory.

Change-Id: Ic9d495255ca48110920cf2df371946ac146dcd72