chore(tazjin/emacs): keep only one workspace on tverskoy's screen
this machine is now sort of permanently stationed in the office, and
thus permanently connected to the big screen.

with this setup, it's comfortable to have it available for a single
workspace (e.g. for videos playing there), but it's too confusing and
unergonomic to use that screen for anything else.

Change-Id: I03556b777c79f68d65d4d8bf1ba1f18982650a8b