feat(3p/cgit): Display depot revisions in log views
depot revisions are not recognised by git as a specific type of
"decoration" (i.e. fall into `DECORATION_NONE`), which means that cgit
skips them by default.

Displaying them on code.tvl.fyi would be quite nice, so to work around
this we check for the `refs/r/` prefix on skipped decorations and
display them (without the `refs/` prefix) on commits.

Change-Id: Ia9cece3876db471f93ab787258be527739af046b