style(3p/cgit): move commit decorations left of subject
the commit decoration layout is designed for situations where very few
commits have decorations, but every depot commit has a revision.

putting these on the left of the commit subject looks a lot more
reasonable. ideally we would actually put them in the table as a
column, but that'd be a much larger change.

Change-Id: I0b44dbef36b64b183d66278d7ab7e4ff17554684
Reviewed-on: https://cl.tvl.fyi/c/depot/+/5892
Tested-by: BuildkiteCI
Reviewed-by: Profpatsch <mail@profpatsch.de>