feat(nix/buildLisp): enable SBCL core compression
SBCL 2.2.6 uses zstd now (which is a bit faster than the old zlib
compression) and, more importantly, nixpkgs has finally enabled
compression support with it, so we can utilise it in depot.

One problem with enabling compression is that also literal paths in the
image can be compressed, preventing Nix from detecting references – this
may cause the Nix store to track dependencies incorrectly.

Change-Id: I901f11583a2b8297631b9b9c0e790dbe4f0c3d56