feat(predlozhnik): wire up the actual предлог/вопрос matching
with this commit the application is sort of functional-ish as
intended.

users can select either cases or prepositions, have the remaining
choices appropriately constrained, and get the right question (i.e.
case) matching when selecting both.

there should be some explanatory translations and it needs to be
prettier, but this kind of does what I wanted.

Change-Id: If5bdaa7f93235dc267bf38cf13c2029f44d68415