feat(tvix/eval): thread codemap through to disassembler
If the disassembler feature is enabled, make sure that an Rc of the
codemap is available through the chunk.

I know that grfn doesn't like feature flags, but I don't see a way to
do this without incurring the runtime cost of that additional
reference count (which affects every single function and thunk).

Change-Id: I700f27ab665a704f73457b19bd2d7efc93828a16