test(tvix/eval): impl Arbitrary for Value + NixAttrs
Impl Arbitrary for the Value and NixAttrs types in the same way we did
for NixString. Value currently only generates non-"internal" values (no
thunks, AttrNotFound, etc.) because that'd require full generation of
tvix bytecode which is a bit more work than I'd like to do now - there's
a `todo!` left in the code for a place where we could allow opting-in to
internal values later.

Change-Id: I07a59e2b1d89cfaa912d4ecebd642caf4ddb040a