refactor(tvix/eval): Encapsulate compilation inside VM
In preparation for allowing compiling additional files given only a VM
(to implement `builtins.import`), encapsulate the construction and
operation of a `Compiler` entirely within a `VM`. This requires the VM
itself to own and prepare both the `globals` hash map and the `Observer`
for the compiler, both of which are passed by-reference when
constructing the compiler.

Change-Id: Ibe4c7ac4fe7be25bbbfc781ee3f88c5c083eef6d