feat(tvix/eval): initial ... workaround ... for builtins in import
Add a hack to pass the set of builtins to `import` itself. This causes
a bootstrap cycle because the `import` builtin needs to be initialised
with the set of globals before being inserted into the globals, which
also must contain itself.

To break out of the cycle this hack wraps the builtins passed to the
compiler in an `Rc` (probably sensible anyways, as they will end up
getting cloned a bunch) and makes use of `unsafe` to mutate the value
inside of the `Rc` without introducing an extra indirection like
RefCell.

As this only happens once (at the time that the builtins are created),
and in a situation where nothing else can access the
under-construction builtins, it is safe enough for now.

However, we should rethink the builtins construction logic in a
different way. This is quite nasty.

Change-Id: I25f8d4d2a7e8472d401c8ba2f4bbf9d86ab2abcb