tvix_tests/eval-okay-concat-strings-{in}valid-utf8: init
cppnix represents strings using std::string, which may be invalid
UTF-8.  In particular, UTF-8-invalid sequences may be spliced
together to produce UTF-8-valid results.  This behavior is closer to
Rust's `[u8]` than its `std::string`.

In the long run we may need to use the `bstr` crate, which was
designed for this situation.  Until then, these tests (both of which
currently fail) can make sure we don't forget to come back to this.

Signed-off-by: Adam Joseph <adam@westernsemico.com>
Change-Id: I107af03a7a1c086bbe3c41ead5d1dd1b5310333d