fix(nix/tag): correct no match check in discr
It uses discrDef internally, but passes `null` as the default tag name,
causing Nix to drop the attribute and return an empty attribute set if
the default case is hit. Consequently we need to check for the empty
attribute set, not `null` to figure out if there was no match found.

We can also test this behavior using `assertThrows` which was introduced
after the tag library was originally written.

Change-Id: I45adb2f9602762dfc867956323fb3f5ae4c8bd1d