wip(wpcarro): Implement "Algorithm W"
I've been wanting to grok Haskell-style type inference for awhile, so instead of
just watching conference talks and reading papers about it, I've decided to
attempt to implement it to more readily test my understanding of it.

See "Wish List" section of the moduledoc for the current state of the project.

Change-Id: I69261202a3d74d55c6e38763d7ddfec73c392465