feat(tvix/eval): Implement builtins.deepSeq
This is done via a new `deepForce` function on Value. Since values can
be cyclical (for example, see the test-case), we need to do some extra
work to avoid RefCell borrow errors if we ever hit a graph cycle: we
introduce a new `try_force` method on Thunk which turns the panic on an
overlapping RefCell borrow into a `false` boolean return, which we can
then use to just stop recursing in try_force.

Finally, since deep_force is conceptually similar to
`VM::force_for_output` (but more suited to usage in eval since it
doesn't clone the values) this removes the latter, replacing it with the
former.

Change-Id: Iefddefcf09fae3b6a4d161a5873febcff54b9157