docs(tvix/eval): upvalues.rs: define "upvalue", comment with_stack
This commit adds a comment for the benefit of non-Lua-users
explaining what an upvalue is.  It also adds a comment explaining
what `with_stack` does, including a brief reminder of nix's
slightly-unusual-but-well-justified handling of this construct.

It was sort of unclear if a with-bound variable "has an upvalue" or
not.  The `upvalues` field doesn't contain the values for those
variables, which would seem to imply "no".  However the `Upvalues`
struct contains `with_stack`.  Is the `with_stack` an upvalue?  If
so, then "yes".  If `with_stack` isn't an upvalue, then why is it
part of `Upvalues`?

Since the terminology comes from Lua, which doesn't have
with-resolved variables, tvix can decide however it likes.  Let's
try to be self-consistent though.  I tried to do this by simply not
defining "an upvalue" but instead defining "upvalues" (plural) and
explaining in a comment that the `upvalues` field contains only the
upvalues of the bound variables.

Signed-off-by: Adam Joseph <adam@westernsemico.com>
Change-Id: If6d0e133292451af906e1c728e34010f99be8c7c