feat(tvix/eval): unify Thunk with Closure (almost)
Still WIP.  Other than accessor boilerplate (of which there is more
than I'd like) the remaining Thunk-specific code is all related to
forcing, and there is essentially no Closure-specific code.

The next step is to turn Closures into known-to-be-forced Thunks
with artiy>0, and the unification will be complete.  If I can
represent known-to-be-forced at the type level, Rust's monomorphizer
will produce code with zero performance penalty relative to the
separate/duplicated implementations.  Still not sure if that's
possible.

Ultimately this may end up not being worth merging, but at least I
feel like I understand the VM now.

Signed-off-by: Adam Joseph <adam@westernsemico.com>
Change-Id: I04131501029772f30e28da8281d864427685097f