feat(tvix/eval): move tests to --feature expected_failures
It is helpful to be able to use the test suite as a regression test:
make a change to the compiler/vm, re-run the tests, and if there are
any failures you know it's your fault.

Right now we can't do that, because the expected-to-fail tests are
mixed in with the expected-to-pass tests.  So we can't use them as a
regression test.

This commit fixes that, by moving all the
currently-failing-for-origin/canon tests out of `nix_tests` and into
`expected_failures`.  You can still run those tests by adding an
extra feature flag.

Eventually tvix will pass all the tests, and this flag will disappear.

Change-Id: Ied606882b9835a7effd7e75bfcf3e5f827e0a2c8