fix(tvix/eval): don't thunk home path resolution to prevent catching
builtins.tryEval shouldn't catch home path resolution errors which it
currently can in tvix. To prevent this cl/7025 proposed a new flag in
the PathResolution error variant which described if the error was
catchable or not. This CL proposes an alternative implementation which
is simpler, but also a bit more implicit.

tryEval relies on thunking to catch errors, so we can bypass this by
simply not thunking the home path resolution. This means that we need to
resolve them unconditionally (and probably in a bit more expensive way
than in the compiler originally), but save on the extra flag in the
error variant.

Change-Id: Ie4b9dc68f62c86d6c7fd5f1c9460c850d97ed1ca